Update geometry-linear-algebraic-ops.md
Small wording fix "that" -> "by which"

Also removed "Computing determinants for larger matrices can be laborious" -- and just left "For larger matrices the intuition is the same."  Computing the determinant using expansion by cofactors is exponential complexity, but in fact computing the determinant is an O(N^3) operation -- one just uses LU factorization and takes the determinant of L and U (one of which will be unity depending on the factorization used).  The determinant of a product is the product of the determinants.  Computing the LU factorization is O(N^3).  In any case, one would not do this by hand.